Move storage of the user's phone number to the session
Originally, the user's phone number was stored in a flash message
between requests (the code request and code verification) but this
was unnecessarily limiting to the verification process. By storing the
number in the session, while a little extra work to manage, it allows
for more flexibility in the process by the user.