Correct Markdown style issues in the README
This change updates README.md after using Markdownlint to ensure that it
meets a quality baseline. I'm not applying all of the default rules
(partly as they don't all apply, and partly as I'm not completely
familiar with Markdownlint, yet. But, I've applied all but two of them.
My intention is to use it ever more to ensure that the documentation is
of as high a level of quality as possible.